+++ wpglobus.php
@@ -6,7 +6,7 @@
  * Description: A WordPress Globalization / Multilingual Plugin. Posts, pages, menus, widgets and even custom fields - in multiple languages!
  * Text Domain: wpglobus
  * Domain Path: /languages/
- * Version: 1.2.9
+ * Version: 1.3.0
  * Author: WPGlobus
  * Author URI: http://www.wpglobus.com/
  * Network: false
@@ -30,7 +30,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'WPGLOBUS_VERSION', '1.2.9' );
+define( 'WPGLOBUS_VERSION', '1.3.0' );
 define( 'WPGLOBUS_PLUGIN_BASENAME', plugin_basename( __FILE__ ) );
 
 /** @todo Get rid of these */
@@ -56,22 +56,30 @@ WPGlobus::Config();
 require_once 'includes/class-wpglobus-filters.php';
 require_once 'includes/wpglobus-controller.php';
 
+if ( defined( 'WPSEO_VERSION' ) ) {
+	require_once 'includes/class-wpglobus-wpseo.php';
+	WPGlobus_WPSEO::controller();
+}
+
 /**
- * Fix multilingual strings in `Customize`
+ * Theme compatibility
+ */
+
+/**
+ * Fix multilingual strings in basic `Customize`
  *
  * @since 1.2.1
  */
 require_once 'includes/class-wpglobus-customize.php';
 WPGlobus_Customize::controller();
 
-if ( defined( 'WPSEO_VERSION' ) ) {
-	require_once 'includes/class-wpglobus-wpseo.php';
-	WPGlobus_WPSEO::controller();
-}
-
-if ( WPGlobus_WP::in_wp_admin() ) {
+/**
+ * Support of theme option panels and customizer
+ * @since 1.3.0
+ */
+if ( WPGlobus_WP::in_wp_admin() && ! WPGlobus_WP::is_admin_doing_ajax() ) {
 	require_once 'includes/admin/class-wpglobus-wp-theme.php';
-	new WPGlobus_WP_Theme();
+	WPGlobus::Config()->WPGlobus_WP_Theme = new WPGlobus_WP_Theme();
 }
 
 require_once 'updater/class-wpglobus-updater.php';